25 november 2024

+++ database/migrations/2024_11_25_061820_create_membership_rewards_table.php
@@ -11,16 +11,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('membership_rewards', function (Blueprint $table) {
-            $table->id(); // Primary key
-            $table->foreignId('membershipLevelId')->constrained('memberships')->onDelete('cascade'); // Foreign key for memberships
-            $table->foreignId('rewardTypeId')->constrained('reward_types')->onDelete('cascade'); // Foreign key for reward types
-            $table->unsignedInteger('bonusPoints')->nullable(); // Bonus points
-            $table->decimal('nominalRequired', 10, 2)->nullable(); // Nominal required
-            $table->string('appliesEvery', 255)->nullable(); // Applies every (e.g., "month", "year")
-            $table->timestamps(); // Created at and updated at timestamps
-            $table->softDeletes(); // Soft delete column
-        });
+        if (!Schema::hasTable('membership_rewards')) {
+            Schema::create('membership_rewards', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('membershipLevelId')->constrained('memberships')->onDelete('cascade');
+                $table->foreignId('rewardTypeId')->constrained('reward_types')->onDelete('cascade');
+                $table->unsignedInteger('bonusPoints')->nullable();
+                $table->decimal('nominalRequired', 10, 2)->nullable();
+                $table->string('appliesEvery', 255)->nullable();
+                $table->timestamps();
+                $table->softDeletes();
+            });
+        }
     }
 
     /**
